requirements.txt
pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
qdrant-client==1.10.1
textract==1.6.5
torch==2.4.0
torchvision==0.19.0
transformers==4.43.2